+++ examples/the-dungeon-puzzlers-lament/src/backgrounds.rs
@@ -1,12 +1,12 @@
 use agb::{
-    display::tiled::{RegularMap, TileFormat, TileSet, TileSetting, VRamManager},
+    display::tiled::{RegularMap, TileFormat, TileSet, VRamManager},
     include_background_gfx,
 };
 
 include_background_gfx!(backgrounds, "1e151b",
-    ui => "maps/ui_tiles.png",
-    level => "maps/level.png",
-    ending => "gfx/ending_page.aseprite",
+    ui => deduplicate "maps/ui_tiles.png",
+    level => deduplicate "maps/level.png",
+    ending => deduplicate "gfx/ending_page.aseprite",
 );
 
 mod tilemaps {
@@ -56,13 +56,13 @@ pub fn load_ending_page(map: &mut RegularMap, vram_manager: &mut VRamManager) {
     for y in 0..20u16 {
         for x in 0..30u16 {
             let tile_pos = y * 30 + x;
-            let tile_setting = TileSetting::new(
-                tile_pos,
-                false,
-                false,
-                backgrounds::ending.palette_assignments[tile_pos as usize],
+
+            map.set_tile(
+                vram_manager,
+                (x, y).into(),
+                &ending_tileset,
+                backgrounds::ending.tile_settings[tile_pos as usize],
             );
-            map.set_tile(vram_manager, (x, y).into(), &ending_tileset, tile_setting);
         }
     }
 }