+++ examples/the-hat-chooses-the-wizard/src/level_display.rs
@@ -3,10 +3,10 @@ use agb::display::{
     HEIGHT, WIDTH,
 };
 
-const LEVEL_START: u16 = 12 * 28;
-const NUMBERS_START: u16 = 12 * 28 + 3;
-const HYPHEN: u16 = 12 * 28 + 11;
-pub const BLANK: u16 = 11 * 28;
+const LEVEL_START: usize = 12 * 28;
+const NUMBERS_START: usize = 12 * 28 + 3;
+const HYPHEN: usize = 12 * 28 + 11;
+pub const BLANK: usize = 11 * 28;
 
 pub fn write_level(
     map: &mut RegularMap,
@@ -14,25 +14,21 @@ pub fn write_level(
     level: u32,
     tileset: &'_ TileSet<'_>,
     vram: &mut VRamManager,
+    tile_settings: &[TileSetting],
 ) {
     for (i, &tile) in [
         LEVEL_START,
         LEVEL_START + 1,
         LEVEL_START + 2,
         BLANK,
-        world as u16 + NUMBERS_START - 1,
+        world as usize + NUMBERS_START - 1,
         HYPHEN,
-        level as u16 + NUMBERS_START - 1,
+        level as usize + NUMBERS_START - 1,
     ]
     .iter()
     .enumerate()
     {
-        map.set_tile(
-            vram,
-            (i as u16, 0).into(),
-            tileset,
-            TileSetting::from_raw(tile),
-        );
+        map.set_tile(vram, (i as u16, 0).into(), tileset, tile_settings[tile]);
     }
 
     map.set_scroll_pos((-(WIDTH / 2 - 7 * 8 / 2) as i16, -(HEIGHT / 2 - 4) as i16).into());